+++ include/algorithm/early-berlekamp-massey.h
@@ -16,7 +16,7 @@ namespace wiedemann {
         long L = 0, m = 1, streak = 1, stop = 0;
         NTL::ZZ_p b = NTL::ZZ_p(1);
 
-        for (long k = 0; k < sequence.size(); ++k) {
+        for (size_t k = 0; k < sequence.size(); ++k) {
             NTL::ZZ_p d = sequence[k];
             for (long i = 1; i <= L; ++i)
                 d += C[i] * sequence[k - i];
